requirements.txt
python-dotenv
python-telegram-bot
./epson-epos
unidecode